include/cpu/amd/mtrr: return uint32_t from get_top_of_mem_below_4gb
The top of memory below 4GB will always fit into 32 bits, so change the
return type accordingly.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/74610
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/include/cpu/amd/mtrr.h
@@ -65,7 +65,7 @@ static __always_inline void wrmsr_amd(unsigned int index, msr_t msr)
 		);
 }
 
-static inline uint64_t get_top_of_mem_below_4gb(void)
+static inline uint32_t get_top_of_mem_below_4gb(void)
 {
 	return rdmsr(TOP_MEM).lo;
 }
